Big improvements to datatable modules
- Clients: Add dropdown filters to datatable.
- Clients: Add JS validation to add client form.
- Clients: Redesign add client form.
- General: Move error handling and notification code out of DB classes.
- General: Add refresh button for datatables.
- General: Fix datatable CSS.

+++ public/modules/users/users.js
@@ -1,8 +1,8 @@
 class Users{
-    overlay = false
+    overlay = true
     
     static async load(arg){
-        let users = await UsersDB.getActiveUsers()
+        let users = await UsersDB.getUsers()
 
         $('#datatable').DataTable({
             data: users,
@@ -12,7 +12,7 @@ class Users{
                 sSearch: '', searchPlaceholder: "Search..." 
             },
             initComplete : function() {
-                $("#datatable_filter").detach().appendTo('#datatableSearch');
+                $("#datatable_filter").detach().appendTo('#datatableSearch')
             },
             columns: [
                 { title: "ID", data: "id", visible: false},
@@ -23,11 +23,21 @@ class Users{
                 { title: "Town", data: "town"},
                 { title: "County", data: "county"},
                 { title: "Eircode", data: "eircode"},
+                { title: "Active", data: "active", visible: false},
                 {mRender: function (data, type, row) {
-                    return `<a href="javascript:Users.deactivateUser('${row.id}')">Deactivate</a>`
+                    if(row.active == true)
+                        return `<a href="javascript:Users.deactivateUser('${row.id}')">Deactivate</a>`
+                    else
+                        return `<a href="javascript:Users.activateUser('${row.id}')">Activate</a>`
+
                 }},
                 {mRender: function (data, type, row) {
-                    return `<a href="javascript:Module.load('UserProfile', '${row.id}')">View Profile</a>`
+                    if(row.role == "Admin")
+                        return ""
+                    else if(row.active == true)
+                        return `<a href="javascript:Module.load('UserProfile', '${row.id}')"> View Profile</a>`
+                    else if(row.active == false)
+                        return `<a href="javascript:Users.deleteUser('${row.id}')">Delete</a>`
                 }},
             ],
         })
@@ -35,49 +45,68 @@ class Users{
         this.listeners()
     }
 
-    // Displays selected user details. 
-    static async viewUserProfile(userId){
-        $('#user-list').hide()
-        $('#user-profile').show() 
-
-        let user = await UsersDB.getUser(userId)
-
-        $('#user-profile-title').html(` ${user.name}'s Profile`)
-        $('#user-profile-id').text(` ${user.id}`)
-        $('#user-profile-name').text(` ${user.name}`)
-        $('#user-profile-mobile').text(` ${user.mobile}`)
-        $('#user-profile-address').text(` ${user.address1}, ${user.address2}, ${user.town}, ${user.county}, ${user.eircode}`)
-
-        this.loadConns(userId)
+    static async activateUser(userId) {
+        await UsersDB.activateUser(userId)
+            .then(() => {
+                Notification.display(1, "User activated")
+            }).catch(error => {
+                Notification.display(2, "Unable to activate user")
+                console.log(error.message)
+            })
+            
+        this.refreshTable()
     }
 
     static async deactivateUser(userId){
-        if(await Prompt.confirm()){
+        if(await Prompt.confirm("This action will remove all clients and visits from this user.")){
 
             await Promise.all([
                 await UsersDB.deactivateUser(userId),
                 await VisitsDB.deleteVisits(userId),
                 await ConnsDB.deleteConns(userId)
-            ])
-
-            //TODO: Don't hardcode. 
-            Notification.display(1, "User deactivated")
+            ]).then(() => {
+                Notification.display(1, "User deactivated")
+            }).catch(error => {
+                Notification.display(2, "Unable to deactivate user")
+                console.log(error.message)
+            })
+            
             this.refreshTable()
         }
+    }
+
+    static async deleteUser(userId){
+        if(await Prompt.confirm()){
+            await UsersDB.deleteUser(userId)
+        }
+
+        this.refreshTable()
     }
     
     // Resets and reloads datatable. 
     static async refreshTable(){
-        let users = await UsersDB.getActiveUsers()
-        let table = $('#datatable').DataTable()
+        startLoad()
 
-        table.clear() 
-        table.rows.add(users)
-        table.draw()
+        UsersDB.getUsers()
+            .then(users => {
+                let table = $('#datatable').DataTable()
+
+                table.clear() 
+                table.rows.add(users)
+                table.draw()
+                endLoad()
+            }).catch(error => {
+                console.log(error.message)
+                Notification.display(2, "Unable to load users")
+                endLoad()
+            })
     }
 
     // Instantiate listeners.
     static async listeners(){
-
+        $('.btn-refresh').click(function(){
+            Animate.rotate(360, '.btn-refresh-icon')
+            Users.refreshTable()
+        })
     }
 }
